Update to 5.8

+++ database/migrations/2017_12_26_155215_create_charity_boxes_table.php
@@ -14,7 +14,7 @@ class CreateCharityBoxesTable extends Migration
     public function up()
     {
         Schema::create('charity_boxes', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             //Numer kolejny puszki
             //$table->integer('boxNumber');
             //Numer puszki ma format numer wolontariusza/numer sztabu
@@ -22,24 +22,24 @@ class CreateCharityBoxesTable extends Migration
             //Identyfikator wolontariusza
             $table->string('collectorIdentifier');
             //ID wolontariusza w bazie
-            $table->integer('collector_id');
+            $table->bigInteger('collector_id');
             //STATUSY
             //Czy puszka została wydana
             $table->boolean('is_given_to_collector')->default(false);
             //Kto wydał puszkę
-            $table->integer('given_to_collector_user_id');
+            $table->bigInteger('given_to_collector_user_id');
             //Data i godzina wydania puszki
             $table->dateTime('time_given');
             //Czy puszka została policzona
             $table->boolean('is_counted')->default(false);
             //Użytkownik rozliczający puszkę
-            $table->integer('counting_user_id')->nullable();
+            $table->bigInteger('counting_user_id')->nullable();
             //Data i godzina przeliczenia puszki
             $table->dateTime('time_counted')->nullable();
             //Czy puszka została zatwierdzona
             $table->boolean('is_confirmed')->default(false);
             //Kto zatwierdził
-            $table->integer('user_confirmed_id')->nullable();
+            $table->bigInteger('user_confirmed_id')->nullable();
             //Czas zatwierdzenia puszki
             $table->dateTime('time_confirmed')->nullable();
             //Ilość monet 1gr
